requirements.txt
iso8601
scheduler
numpy
pyyaml
tensorflow
python-telegram-bot
pandas
scikit-learn
requests
matplotlib